requirements.txt
discord.py==2.4.0
gTTS==2.5.4
Pillow==11.2.1
psutil==6.1.1
Requests==2.32.3
yt_dlp==2025.4.30
